Referenced unions that include unreferenced types breaks the type system. Fix this by removing the unreferenced type from the union.
I *wanted* to fix this by including the unreferenced type as an empty type. But graphql.js doesn't allow empty types... it's not clear whether this is required by the GraphQL specification or not, but it really doesn't matter because graphql.js is so important in the ecosystem.

This probably leaves an unresolved bug: if you reference a union but don't reference any of its underlying types, this will probably break things. That seems really unlikely because why would you do that? Just to get the __typename? But it's still worth making that error explicit in the future.